backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
boto3==1.34.19
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
ldap3==2.9.1
bcrypt==4.1.2
Pillow==10.2.0
